# openapi.yaml
openapi: 3.0.0
info:
  title: Binar Academy Post API
  version: 1.0.0
servers:
  - url: https://localhost:8000
    description: Development server
  - url: https://staging.com
    description: Staging server
  - url: https://production.com
    description: Production server
...
paths:
  /api/v1/cars:
    get:
      tags:
        - Car
      summary: Get all data for user
      description: To retrive the list of posts
      responses:
        '200':
          description: If data successfully retrived
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  data:
                    type: object
                    properties:
                      posts:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              format: int64
                              example: 1
                            name:
                              type: string
                              example: Honda
                            price:
                              type: integer
                              example: 20000
                            size:
                              type: string
                              example: large
                            image:
                              type: string
                              example: Test image
                            createdBy:
                              type: integer
                              example: 1
                            deletedBy:
                              type: integer
                              example: 2
                            updatedBy:
                              type: integer
                              example: 1
                            available:
                              type: boolean
                              example: true or false
                            createdAt:
                              type: string
                              example: '2022-10-12T13:25:21.995Z'
                            updatedAt:
                              type: string
                              example: '2022-10-12T13:25:21.995Z'
                      count:
                        type: integer
                        format: int64
                        example: 10
        '404':
          description: If failed to retrive the data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: FAIL

    post:
        tags:
          - Car
        summary: Create cars data
        requestBody:
          content:
            application/json:
              schema:
                          properties:
                            name:
                              type: string
                              example: Honda
                            size:
                              type: string
                              example: medium
                            price:
                              type: number
                              example: 10000
                            image:
                              type: string
                              example: 'Honda.jpg'
          required: true
        responses:
          '200':
            description: A post was created
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: Created
                    data:
                      type: object
                      properties:
                        posts:
                          type: array
                          items:
                            properties:
                              id:
                                type: integer
                                format: int64
                                example: 3
                              name:
                                type: string
                                example: Honda
                              size:
                                type: string
                                example: Medium
                              createdBy:
                                type: integer
                                example: '1'
                              image:
                                type: string
                                example: "http://res.cloudinary.com/dptgh7efj/image/upload/v1666344280/samples/fxyfzjgujxlbs8md6wtc.jpg"  
                              updatedAt:
                                type: string
                                example: '2022-10-21T09:24:41.017Z'
                              createdAt:
                                type: string
                                example: '2022-10-21T09:24:41.017Z'
                              updatedBy:
                                type: string
                                example: null
                              deletedBy:
                                type: string
                                example: null
                                
          '401':
            description: If member tried to put cars data
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: Unauthorized
                    message:
                      type: string
                      example: Danger area, not user access zone
  /api/v1/cars/{id}:
    get:
      tags:
        - Car
      summary: Get data by id
      description: To retrive data by id
      parameters:
      - name: id
        in: path
        description: ID of the post
        required: true
        schema:
          type: string
      responses:
        '200':
          description: If data successfully retrived
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  data:
                    type: object
                    properties:
                      posts:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              format: int64
                              example: 1
                            title:
                              type: string
                              example: Test Title
                            body:
                              type: string
                              example: Test Body
                            createdAt:
                              type: string
                              example: '2022-10-12T13:25:21.995Z'
                            updatedAt:
                              type: string
                              example: '2022-10-12T13:25:21.995Z'
                      count:
                        type: integer
                        format: int64
                        example: 10
        '400':
          description: If failed to retrive the data
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: FAIL
                  message:
                    type: string
                    example: Bad request
    
    put:
      tags:
        - Car
      summary: Put data by id
      description: To retrive the list of posts
      parameters:
      - name: id
        in: path
        description: ID of the Car
        required: true
        schema:
          properties:
           id:
            type: object
            example: 3
      requestBody:
        content:
          application/json:
            schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Ferrari
                  price:
                    type: number
                    example: 3200000
        required: true
      responses:
        '200':
          description: If data successfully retrived
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
        '401':
          description: If member tried to put cars data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Danger area, not user access zone
                    
    delete:
      tags:
        - Car
      summary: Delete data by id
      description: To delete data by id
      parameters:
      - name: id
        in: path
        description: ID of the post
        required: true
        schema:
          properties:
            id:
              type: object
              example: 3
      responses:
        '204':
          description: no content
...